move test case to failpoint directory
