+++ tests/integrations/server/kv_service.rs
@@ -3632,94 +3632,3 @@ fn test_check_cluster_id_for_batch_cmds() {
         }
     }
 }
-
-// #[test_case(test_raftstore::must_new_cluster_and_kv_client)]
-// #[test_case(test_raftstore_v2::must_new_cluster_and_kv_client)]
-#[test]
-fn test_scan_locks_with_in_progress_transfer_leader() {
-    let (mut cluster, _, mut ctx) = must_new_cluster_and_kv_client_mul(3);
-    cluster.pd_client.disable_default_operator();
-
-    cluster.must_transfer_leader(1, new_peer(1, 1));
-    let leader_peer = cluster.leader_of_region(1).unwrap();
-    ctx.set_peer(leader_peer.clone());
-    let k1 = b"k1";
-    let k2 = b"k2";
-    let leader_region = cluster.get_region(k1);
-    ctx.set_region_epoch(leader_region.get_region_epoch().clone());
-    let env = Arc::new(Environment::new(1));
-    let channel =
-        ChannelBuilder::new(env).connect(&cluster.sim.rl().get_addr(leader_peer.get_store_id()));
-    let client = TikvClient::new(channel);
-
-    // Create both pessimistic locks.
-    let start_ts = 10;
-    must_kv_pessimistic_lock(&client, ctx.clone(), k1.to_vec(), start_ts);
-    must_kv_pessimistic_lock(&client, ctx.clone(), k2.to_vec(), start_ts);
-
-    // Ensure the pessimistic locks are written to the memory but not the storage.
-    let engine = cluster.get_engine(leader_peer.get_store_id());
-    let cf_res = engine
-        .get_value_cf(
-            CF_LOCK,
-            keys::data_key(Key::from_raw(k1).as_encoded()).as_slice(),
-        )
-        .unwrap();
-    assert!(cf_res.is_none());
-    let cf_res = engine
-        .get_value_cf(
-            CF_LOCK,
-            keys::data_key(Key::from_raw(k2).as_encoded()).as_slice(),
-        )
-        .unwrap();
-    assert!(cf_res.is_none());
-
-    let mut scan_lock_req = ScanLockRequest::default();
-    scan_lock_req.set_context(ctx.clone());
-    scan_lock_req.max_version = start_ts + 10;
-    scan_lock_req.limit = 256;
-    let scan_lock_resp = client.kv_scan_lock(&scan_lock_req.clone()).unwrap();
-    assert!(!scan_lock_resp.has_region_error());
-    assert_eq!(scan_lock_resp.get_locks().len(), 2);
-    assert_eq!(scan_lock_resp.locks.to_vec()[0].lock_version, start_ts);
-    assert_eq!(scan_lock_resp.locks.to_vec()[0].key, k1);
-    assert_eq!(scan_lock_resp.locks.to_vec()[1].lock_version, start_ts);
-    assert_eq!(scan_lock_resp.locks.to_vec()[1].key, k2);
-
-    // Propose the transfer leader command but only trigger proposing pessimistic
-    // locks.
-    fail::cfg(
-        "finish_proposing_transfer_cmd_after_proposing_locks",
-        "return",
-    )
-    .unwrap();
-    let _ = cluster.try_transfer_leader_with_timeout(1, new_peer(2, 2), Duration::from_secs(1));
-    let timer = tikv_util::time::Instant::now();
-    let timeout = Duration::from_secs(5);
-    loop {
-        let cf_res = engine
-            .get_value_cf(
-                CF_LOCK,
-                keys::data_key(Key::from_raw(k1).as_encoded()).as_slice(),
-            )
-            .unwrap();
-        if cf_res.is_some() {
-            break;
-        }
-        std::thread::sleep(Duration::from_secs(1));
-        if timer.saturating_elapsed() >= timeout {
-            assert!(cf_res.is_some());
-            break;
-        }
-    }
-    fail::remove("finish_proposing_transfer_cmd_after_proposing_locks");
-
-    // Verify there should be no duplicate locks returned.
-    let scan_lock_resp = client.kv_scan_lock(&scan_lock_req.clone()).unwrap();
-    assert!(!scan_lock_resp.has_region_error());
-    assert_eq!(scan_lock_resp.locks.len(), 2);
-    assert_eq!(scan_lock_resp.locks.to_vec()[0].lock_version, start_ts);
-    assert_eq!(scan_lock_resp.locks.to_vec()[0].key, k1);
-    assert_eq!(scan_lock_resp.locks.to_vec()[1].lock_version, start_ts);
-    assert_eq!(scan_lock_resp.locks.to_vec()[1].key, k2);
-}
